Refactored PropertiesSystemProperties into PropertiesSystemProperties andStringSourceSystemProperties, a base class fore EZ config plugins.
